restructure rpc_state_reader crate
