+++ rpc_state_reader/Cargo.toml
@@ -3,7 +3,8 @@ name = "rpc_state_reader"
 version = "0.4.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+starknet_in_rust = []
 
 [dependencies]
 ureq = { version = "2.7.1", features = ["json"] }
